requirements.txt
django>=2.2.6,<2.3.0
djangorestframework>=3.10.3,<3.11.0
django-cors-headers>=3.2.1,<3.3.0

# Code linting tests
flake8>=3.7.8,<3.8.0

# Database
psycopg2>=2.8.3,<2.9.0

gunicorn>=19.9.0,<=29.0.0